requirements.txt
Flask~=3.1.1
Flask-SocketIO~=5.5.1
gevent~=25.5.1
HomeAssistant-API~=5.0.0
gunicorn~=23.0.0
requests~=2.32.3
python-dotenv~=1.1.0